+++ src/main/java/gift/dto/ProductRequestDto.java
@@ -1,11 +1,13 @@
 package gift.dto;
 
+import java.math.BigDecimal;
+
 public class ProductRequestDto {
     private final String name;
-    private final int price;
+    private final BigDecimal price;
     private final String imgUrl;
 
-    public ProductRequestDto(String name, int price, String imgUrl) {
+    public ProductRequestDto(String name, BigDecimal price, String imgUrl) {
         this.name = name;
         this.price = price;
         this.imgUrl = imgUrl;
@@ -15,7 +17,7 @@ public class ProductRequestDto {
         return name;
     }
 
-    public int getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
